wallets now also use the stat post update logic
